rename: auth request,response 파일 구조 수정
