+++ src/main/java/com/study/modoos/auth/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.study.modoos.auth.dto;
+package com.study.modoos.auth.request;
 
 import lombok.Getter;
